fix: Exception Error
Exception Error 수정

+++ src/main/java/com/ll/medium/global/exception/PostNotFoundException.java
@@ -2,13 +2,13 @@ package com.ll.medium.global.exception;
 
 public class PostNotFoundException extends RuntimeException{
 
-    private final Integer postId;
-    public PostNotFoundException(Integer postId) {
+    private final Long postId;
+    public PostNotFoundException(Long postId) {
         super("Post not found with id:" + postId);
         this.postId = postId;
     }
 
-    public Integer getPostId() {
+    public Long getPostId() {
         return postId;
     }
 }
